+++ android/src/main/kotlin/com/weplenish/flutter_bt/ConnectedBtDevice.kt
@@ -3,26 +3,9 @@ package com.weplenish.flutter_bt
 import android.bluetooth.BluetoothAdapter
 import android.bluetooth.BluetoothDevice
 import android.os.Build
-import android.os.Handler
-import android.os.Looper
 import androidx.annotation.RequiresApi
-import io.flutter.plugin.common.BinaryMessenger
-import io.flutter.plugin.common.EventChannel
-import io.flutter.plugin.common.MethodCall
-import io.flutter.plugin.common.MethodChannel
 
-class ConnectedBtDevice(private val device: BluetoothDevice, private val binaryMessenger: BinaryMessenger) : MethodChannel.MethodCallHandler {
-    private val bluetoothAdapter: BluetoothAdapter? by lazy(LazyThreadSafetyMode.NONE) {
-        BluetoothAdapter.getDefaultAdapter()
-    }
-
-    private val pairedDevice: BluetoothDevice? by lazy(LazyThreadSafetyMode.NONE) {
-        bluetoothAdapter?.bondedDevices?.firstOrNull { it.address == device.address }
-    }
-
-    private val methodChannel: MethodChannel = MethodChannel(binaryMessenger, "com.weplenish.flutter_bt.${device.address}")
-    private val eventChannel: EventChannel = EventChannel(binaryMessenger, "com.weplenish.flutter_bt.event.${device.address}")
-
+class ConnectedBtDevice {
     companion object {
         const val GET_UUIDS = "GET_UUIDS"
         const val DEVICE_INFO = "DEVICE_INFO"
@@ -37,13 +20,17 @@ class ConnectedBtDevice(private val device: BluetoothDevice, private val binaryM
         const val DEVICE_CLASS = "deviceClass"
         const val MAJOR_DEVICE_CLASS = "majorDeviceClass"
 
+        fun deviceUuidToList(device: BluetoothDevice) : List<String>? {
+            return device.uuids?.map { uuidParcel -> uuidParcel.uuid.toString() }
+        }
+
         @RequiresApi(Build.VERSION_CODES.JELLY_BEAN_MR2)
         fun deviceToMap(device: BluetoothDevice?) : Map<String, Any?>? {
             return device?.let {
                 mapOf(
                         ADDRESS to it.address,
                         NAME to it.name,
-                        UUIDS to it.uuids?.map { uuidParcel -> uuidParcel.uuid.toString() },
+                        UUIDS to deviceUuidToList(it),
                         BOND_STATE to it.bondState,
                         TYPE to it.type,
                         BLUETOOTH_CLASS to it.bluetoothClass.let { btClass ->
@@ -53,18 +40,6 @@ class ConnectedBtDevice(private val device: BluetoothDevice, private val binaryM
                             )
                         }
                 )
-            }
-        }
-    }
-
-    @RequiresApi(Build.VERSION_CODES.JELLY_BEAN_MR2)
-    override fun onMethodCall(call: MethodCall, result: MethodChannel.Result) {
-        when(call.method){
-            GET_UUIDS -> {
-                result.success(pairedDevice?.uuids?.map { uuidParcel -> uuidParcel.uuid.toString() })
-            }
-            else -> {
-                result.notImplemented()
             }
         }
     }